requirements.txt
Django==2.2
django-leaflet==0.27.1
django-restframework==0.0.1
psycopg2==2.8.6
pytz==2020.4
sqlparse==0.4.1
